+++ .github/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 * Adds a new algorithm for hafnians of matrices with low rank. [#166](https://github.com/XanaduAI/thewalrus/pull/166)
 
+* Adds a function to calculate the fidelity between two Gaussian quantum states. [#169](https://github.com/XanaduAI/thewalrus/pull/169)
+
+* Adds a new module, `thewalrus.random`, to generate random unitary, symplectic and covariance matrices. [#169](https://github.com/XanaduAI/thewalrus/pull/169)
+
 ### Improvements
 
 * Removes support for Python 3.5. [#163](https://github.com/XanaduAI/thewalrus/pull/163)